Adding support for edit Position and delete

+++ CommonLayer/Entities/Position.cs
@@ -10,7 +10,7 @@ namespace CommonLayer.Entities
     {
         public int Id { get; set; }
         public string PositionName { get; set; }
-        public int BaseSalary { get; set; }
+        public decimal BaseSalary { get; set; }
         public int DepartmentId { get; set; }
     }
 }
